requirements.txt
alembic==0.8.2
boto==2.38.0
decorator==4.0.2
Flask==0.10.1
Flask-Admin==1.1.0
Flask-Assets==0.10
Flask-Migrate==1.5.1
Flask-S3==0.1.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.0
Frozen-Flask==0.11
gnureadline==6.3.3
ipython==3.1.0
itsdangerous==0.24
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
pbr==1.7.0
PySmbClient==0.1.3
python-editor==0.4
six==1.9.0
SQLAlchemy==1.0.5
sqlparse==0.1.16
Tempita==0.5.2
webassets==0.10.1
Werkzeug==0.10.4
wheel==0.24.0
WTForms==2.0.2
